requirements.txt
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
whitenoise==5.2.0

django-plotly-dash==1.6.3

redis==3.5.3
channels-redis==3.2.0
daphne==3.0.2
django-redis==4.12.1
channels-redis==3.2.0

pandas==1.2.4
openpyxl==3.0.7